% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRIT.R
\name{gRIT}
\alias{gRIT}
\title{Generalized random intersection trees}
\usage{
gRIT(x, y, rand.forest = NULL, read.forest = NULL,
  rit.param = list(depth = 5, ntree = 500, nchild = 2, class.id = 1,
  min.nd = 1, class.cut = NULL), varnames.grp = colnames(x),
  weights = rep(1, nrow(x)), signed = TRUE, oob.importance = TRUE,
  ints.idx.eval = NULL, ints.eval = NULL, n.core = 1)
}
\arguments{
\item{x}{numeric feature matrix}

\item{y}{response vector. If factor, classification is assumed.}

\item{rand.forest}{an object of class randomForest. Required if read.forest
is NULL.}

\item{read.forest}{output of readForest. Required if rand.forest is NULL.}

\item{rit.param}{named list specifying RIT parameters. Entries include
\code{depth}: depths of RITs, \code{ntree}: number of RITs, \code{nchild}:
number of child nodes for each RIT, \code{class.id}: 0-1 indicating which
leaf nodes RIT should be run over, \code{min.nd}: minimum node size to run
RIT over, \code{class.cut}: threshold for converting leaf nodes in
regression to binary classes.}

\item{varnames.grp}{grouping "hyper-features" for RIT search. Features with
the same name will be treated as identical for interaction search.}

\item{weights}{numeric weight for each observation. Leaf nodes will be
sampled for RIT with probability proprtional to the total weight of
observations they contain.}

\item{signed}{if TRUE, signed interactions will be returned}

\item{oob.importance}{if TRUE, importance measures are evaluated on OOB
samples.}

\item{ints.idx.eval}{like \code{ints.eval}, but specifies the indice of the
interactions instead of their names. Intended for internal use only.}

\item{ints.eval}{interactions to evaluate. If specified, importance metrics
will be evaluated for these interactions instead of those recovered by RIT.}

\item{n.core}{number of cores to use. If -1, all available cores are used.}
}
\value{
a data table containing the recovered interactions and importance
 scores.
}
\description{
Run RIT across decision paths of a fitted random forest.
}
