% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.randomForest}
\alias{importance.randomForest}
\title{Importance method for random forest}
\usage{
\method{importance}{randomForest}(x, type = NULL, class = NULL,
  scale = TRUE, ...)
}
\description{
Importance method for random forest
}
