% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactPredict.R
\name{interactPredict}
\alias{interactPredict}
\title{Predict interaction}
\usage{
interactPredict(x, int, read.forest, varnames = NULL, min.nd = 1)
}
\arguments{
\item{x}{numeric feature matrix}

\item{int}{a signed interaction. Formatted as 'X1+_X2+_X3-_...'}

\item{read.forest}{output of readForest.}

\item{varnames}{grouping "hyper-features" for RIT search. Features with
the same name will be treated as identical for interaction search.}

\item{min.nd}{minimum leaf node size to extract decision rules from.}
}
\value{
a numeric vector of length nrow(x), entries indicating a predicted
 response for the corresponding observation. Predictions are generated from
 random forest decision rules using only the features in int.
}
\description{
Generate predictions from random forest decision rules corresponding to a
signed interaction.
}
