% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfacePlot.R
\name{plotInt}
\alias{plotInt}
\title{Plot interaction}
\usage{
plotInt(x, y, int, read.forest, varnames = NULL, qcut = 0.5,
  col.pal = c("#1c3f66", "#306aab", "#6e96c4", "#ffb003", "#ff8300"),
  xlab = NULL, ylab = NULL, zlab = NULL, range.col = NULL,
  z.range = c(0, 1), grid.size = 50, min.surface = 100, min.nd = 5,
  pred.prob = FALSE, filt.rule = TRUE, nc = 2, main = NULL)
}
\arguments{
\item{x}{numeric feature matrix, with replicate features grouped}

\item{y}{response vector}

\item{int}{signed interaction to plot. Formatted as 'X1+_X2+_X3-_...'}

\item{read.forest}{output of readForest}

\item{varnames}{character vector indicating feature names. If NULL,
colnames(x) are used as feature names.}

\item{col.pal}{color palette for response surfaces}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{zlab}{z-axis label}

\item{range.col}{range of response values for color palette}

\item{z.range}{z-axiz range}

\item{min.surface}{minimum number of observations required to generate a
response surface.}

\item{min.nd}{minimum leaf node size to extract decision rules from.}

\item{main}{plot title for response surfaces}

\item{qcut:}{quantile to define low/high levels of additional features beyond
order-2 interations. Thresholds will generated using the specified quantile
of random forest thresholds for corresponding features.}

\item{grid.surface}{size of grid to generate response surfaces over.}

\item{pred.prob:}{if TRUE, z-axis indicates predicted probability from the
random forest. If false, z-axis indicates distribution of responses y}
}
\description{
Generate response surface plots for a given interaction.
}
