% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.formula.R
\name{randomForest.formula}
\alias{randomForest.formula}
\title{Random forest formula}
\usage{
\method{randomForest}{formula}(formula, data = NULL, ..., subset,
  na.action = na.fail)
}
\description{
Random forest formula
}
