% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityScore.R
\name{stabilityScore}
\alias{stabilityScore}
\title{Stability score}
\usage{
stabilityScore(x, y, ntree = 500, mtry.select.prob = rep(1, ncol(x)),
  ints.eval = NULL, rit.param = list(depth = 5, ntree = 500, nchild =
  2, class.id = 1, min.nd = 1, class.cut = NULL), varnames.grp = NULL,
  n.bootstrap = 1, bs.sample = NULL, weights = rep(1, nrow(x)),
  signed = TRUE, n.core = 1, ...)
}
\arguments{
\item{x}{numeric feature matrix}

\item{y}{response vector. If factor, classification is assumed.}

\item{ntree}{number of random forest trees.}

\item{mtry.select.prob}{feature weights for first iteration. Defaults to
equal weights}

\item{ints.eval}{interactions to evaluate. If specified, importance metrics
will be evaluated for these interactions instead of those recovered by RIT.}

\item{rit.param}{named list specifying RIT parameters. Entries include
\code{depth}: depths of RITs, \code{ntree}: number of RITs, \code{nchild}:
number of child nodes for each RIT, \code{class.id}: 0-1 indicating which
leaf nodes RIT should be run over, \code{min.nd}: minimum node size to run
RIT over, \code{class.cut}: threshold for converting leaf nodes in
regression to binary classes.}

\item{n.bootstrap}{number of bootstrap samples to calculate stability scores.}

\item{bs.sample}{list of observation indices to use for bootstrap samples. If
NULL, iRF will take standard bootstrap samples of observations.}

\item{weights}{numeric weight for each observation. Leaf nodes will be
sampled for RIT with probability proprtional to the total weight of
observations they contain.}

\item{signed}{if TRUE, signed interactions will be returned}

\item{n.core}{number of cores to use. If -1, all available cores are used.}

\item{...}{additional arguments passed to iRF::randomForest.}
}
\value{
a data table containing the recovered interactions and importance
 scores.
}
\description{
Run outer layer bootstrap stability analysis
}
