% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRF.R
\name{iRF}
\alias{iRF}
\title{Iterative random forests (iRF)}
\usage{
iRF(x, y, xtest = NULL, ytest = NULL, n.iter = 5, ntree = 500,
  mtry.select.prob = rep(1, ncol(x)), iter.return = n.iter,
  int.return = NULL, select.iter = FALSE, rit.param = list(depth = 5,
  ntree = 500, nchild = 2, class.id = 1, min.nd = 1, class.cut = NULL),
  varnames.grp = colnames(x), n.bootstrap = 1, bs.sample = NULL,
  weights = rep(1, nrow(x)), signed = TRUE, oob.importance = TRUE,
  type = "randomForest", verbose = TRUE, n.core = 1,
  interactions.return = NULL, wt.pred.accuracy = NULL, ...)
}
\arguments{
\item{x}{numeric feature matrix.}

\item{y}{response vector. If factor, classification is assumed.}

\item{xtest}{numeric feature matrix for test set.}

\item{ytest}{response vector for test set.}

\item{n.iter}{number of iterations to run.}

\item{ntree}{number of random forest trees.}

\item{mtry.select.prob}{feature weights for first iteration. Defaults to
equal weights}

\item{iter.return}{which iterations should the RF be returned for.
Defaults to iteration with highest OOB accuracy.}

\item{int.return}{which iterations should interacitons be returned for.}

\item{select.iter}{if TRUE, returns interactions from iteration with highest
OOB accuracy.}

\item{rit.param}{named list specifying RIT parameters. Entries include
\code{depth}: depths of RITs, \code{ntree}: number of RITs, \code{nchild}:
number of child nodes for each RIT, \code{class.id}: 0-1 indicating which
leaf nodes RIT should be run over, \code{min.nd}: minimum node size to run
RIT over, \code{class.cut}: threshold for converting leaf nodes in
regression to binary classes.}

\item{varnames.grp}{grouping "hyper-features" for RIT search. Features with 
the same name will be treated as identical for interaction search.}

\item{n.bootstrap}{number of bootstrap samples to calculate stability
scores.}

\item{bs.sample}{list of observation indices to use for bootstrap samples.
If NULL, iRF will take standard bootstrap samples of observations.}

\item{weights}{numeric weight for each observation. Leaf nodes will be
sampled for RIT with probability proprtional to the total weight of
observations they contain.}

\item{signed}{if TRUE, signed interactions will be returned.}

\item{oob.importance}{if TRUE, importance measures are evaluated on OOB
samples.}

\item{verbose}{if TRUE, display progress of iRF fit.}

\item{n.core}{number of cores to use. If -1, all available cores are used.}

\item{...}{additional arguments passed to iRF::randomForest.}
}
\value{
A list containing the following entries:
\itemize{
   \item{rf.list}{a list of randomForest objects}
   \item{interaction}{a data table containing recovered interactions and
     importance scores}
   \item{selected.iter}{iterations returned by iRF}
   \item{weights}{feature weights used to fit each entry of rf.list}
 }
}
\description{
Iteratively grow feature weighted random forests and search for prevalent
interactions on decision paths.
}
