% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForest.R
\name{readForest}
\alias{readForest}
\title{Read forest}
\usage{
readForest(rand.forest, x, return.node.feature = TRUE,
  return.node.obs = FALSE, varnames.grp = NULL, get.split = FALSE,
  first.split = TRUE, n.core = 1)
}
\arguments{
\item{rand.forest}{an object of class randomForest.}

\item{x}{numeric feature matrix}

\item{return.node.feature}{if True, will return sparse matrix indicating
features used on each decision path of the rand.forest}

\item{return.node.obs}{if True, will return sparse matrix indicating
observations in x that fall in each leaf node of rand.forest.}

\item{varnames.grp}{grouping "hyper-features" for RIT search. Features with
the same name will be treated as identical for interaction search.}

\item{get.split}{if True, node feature matrix will indicate splitting
threshold for each selected feature.}

\item{first.split}{if True, splitting threshold will only be evaluated for
the first time a feature is selected.}

\item{n.core}{number of cores to use. If -1, all available cores are used.}
}
\value{
a list containing the follosing entries
\itemize{
   \item{tree.info}{data frame of metadata for each leaf node in rand.forest}
   \item{node.feature}{optional sparse matrix indicating feature usage on
     each decision path}
   \item{node.obs}{optional sparse matrix indicating observations appearing
     in each leaf node}
 }
}
\description{
Read out metadata from random forest decision paths
}
